Make partition function to be generic
